+++ server/src/main/java/com/kalyvianakis/estiator/io/service/IUserService.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface IUserService {
     public User save(User user);
     public List<User> get();
-    public User get(Integer id) throws ResourceNotFoundException;
+    public User get(Long id) throws ResourceNotFoundException;
     public User getOneByEmail(String email) throws ResourceNotFoundException;
     public List<User> getRegistered();
-    public void delete(int id);
-    public Boolean exists(int id);
-    public Boolean notExists(int id);
+    public void delete(Long id);
+    public Boolean exists(Long id);
+    public Boolean notExists(Long id);
 }
